test lint| Simplify pylint invocation in tox
The key idea behind this commit is to help move towards higher quality
code that is the same standard as the main module. The way its
structured means the quality should not decrease overall.

We achieve this by making the test-dirs that need testing into modules
and ignoring directories that contain test-data and fixtures. The key
difference here is that linting is now opt-out rather than opt-in,
meaning all new files added are put under lint-test.